Implement bandwidth monitoring.
We now have all of the functionality available to do this,
so finish implementing it.  Note that in order to do this,
I changed the TopicInfo class from a "is-a" ROSTopicHz to
a "has-a" ROSTopicHz, and also added a "has-a" ROSTopicBandwidth.
With all of that in place, it is pretty simple to fetch
the bandwidth data and display it.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>